Animations are now measured explicitely in seconds

+++ Animations/src/main/java/animations/animation/AbstractAnimation.java
@@ -5,8 +5,17 @@ import javax.swing.JComponent;
 
 /**
  *
- * @author Matt
+ * @author Matt Crow
  */
 public abstract class AbstractAnimation {
-    public abstract void paint(JComponent component, Graphics g, int t);
+    /**
+     * This method is invoked every frame by an AnimationTimer, which renders
+     * this Animation on a given Graphics context.
+     * 
+     * @param component
+     * @param g
+     * @param t seconds elapsed since the start of the animation. 
+     * Ranges from [0, infinity)
+     */
+    public abstract void paint(JComponent component, Graphics g, double t);
 }
